create constants.js for basket and auth module

+++ src/modules/basket/basket.controller.js
@@ -1,3 +1,5 @@
+const statusCode = require("../../core/statusCode.enum");
+
 const BasketError = require("./basket.error");
 const BaskerHelper = require("./basket.helper");
 const BasketService = require("./basket.service");
@@ -11,26 +13,29 @@ class BasketController {
     const user = await BasketService.findUser(req.session.user._id);
     const basket = await BasketService.addProduct(product, user);
 
-    BaskerHelper.responseObject(basket, 200, res);
+    BaskerHelper.responseObject(basket, statusCode.ok, res);
   }
 
   async removeFromBasket(req, res) {
+    if (!req.session.user) return next(BasketError.hasNotSession());
+
     const { _id } = req.session.user;
     const { id } = req.params;
 
     const products = await BasketService.removeProduct(id, _id);
     const price = BaskerHelper.computePrice(products);
     const basket = { products, price };
-    BaskerHelper.responseObject(basket, 200, res);
+    BaskerHelper.responseObject(basket, statusCode.ok, res);
   }
 
-  async getBasket(req, res) {
+  async getBasket(req, res, next) {
+    if (!req.session.user) return next(BasketError.hasNotSession());
     const user = await BasketService.findUser(req.session.user._id);
     const userBasket = await user.populate("basket.items.productId");
 
     const basket = BaskerHelper.mapBasketItems(userBasket.basket);
 
-    BaskerHelper.responseObjWithPrice(basket, 200, res);
+    BaskerHelper.responseObjWithPrice(basket, statusCode.ok, res);
   }
 }
